Expression language proof of concept using Parser Combinator
This is an initial commit of tests with Parser Combinator.  After talking to
Doug I realized the expressions would be different than I expected.  There
are changes I will be making to grammar rules and what gets returned.
I'm thinking to always return a Seq[String], where an expression like
pairs.control.bam would return a list of each bam for every pair's control,
and perhaps a query that ends at an entity returns the json for the entity
model object (this case would be useful for using this expression language
outside a task config or for debugging purposes, like seeing attributes
on an entity and building expressions iteratively).  There is some VERY UGLY
code from when I was trying to allow Vertexes to be returned back, I'm ditching
that in favor of Seq[String] for now.