[CA-1460] Fix Failure to Delete Errored V1 Billing Projects (#1541)
RR: https://broadworkbench.atlassian.net/browse/CA-1460
Test that google project that backs the billing project exists before trying to delete it.

When a user creates a v1 billing account, rawls creates the resource for the google project in sam BEFORE it creates the google project proper. If the project creation fails, we delete the project from rawl’s database but we never tell sam.When listing billing projects, we ask sam for all resources with resource type billing project. Since it doesn’t know that the billing project and google project don’t exist, it returns us all the resources AS IF it existed.

TLDR; we can’t trust that a resource exists just because sam says it does. This is seems reasonable because a user could stab us in the back and delete these resources in google and we’d be in the same situation.