[CA-1628] Create Storage Transfer Job code (#1585)
RR: https://broadworkbench.atlassian.net/browse/CA-1628

In this change, @s-rubenstein and I have added functionality to create storage transfer servive jobs to transfer the contents of one workspace bucket to another.
To do this, we've added a new table called "PPW_STORAGE_TRANSFER_SERVICE_JOB" that tracks STS jobs related to workspace migration attempts. The idea is we create a job and write into the database and then poll sts for updates on each job (we could write a bespoke actor for this). When the job is done, we move to the next stage of migration.

This change seems a little large and most of the noise has been generated from moving things around. In this change:

I decided that the "WorkspaceMigrationMonitor" was getting too large and moved the database-related types and functions into their own files, private to the newly created "migration" package.
I dropped the "V1" prefix because it's unlikely that we'll still be migrating v1 workspaces by the time v3 comes along
I eliminated the use of behavioural inheritance in these types and hid the specifics of table "truncation" in the monitor.