[CA-1702][CA-1704][CA-1708] Project-Per-Workspace Workspace Migration Actor (#1605)
RR: https://broadworkbench.atlassian.net/browse/CA-1702, https://broadworkbench.atlassian.net/browse/CA-1704, https://broadworkbench.atlassian.net/browse/CA-1708

Add functionality to poll the google storage transfer service for status updates of running jobs. To do this, we
- peek the database for the first non-finished job that was updated last (ie the longest time ago)
- touch that database reocrd so that future peeks will return another record first
- fetch the operations of the job from the storage transfer service and combine them into an Outcome write the outcome of the job

If the job finished, we
- update the status of the migration that job relates to
- set the relevent timestamp if the job succeeded or
- fail the migration if the job failed (we can consider retries another time)

Update PPW_STORAGE_TRANSFER_SERVICE_JOB to have an UPDATED column that automagically updates when you touch the record.

I've included a rough implementation of the Akka Actor as I used this to inform my implementation for polling.
The way it works is this: the system scheduler messages the actor to:
- run the migration pipeline every 10s
- update an active transfer job every 5s (two jobs are created per migration)
- We can make these values configurable at another time,

To make this more testable, the migration pipeline runs in reverse. This allows me to set-up a record in the database, run migrate and make sure the correct step picked up the record and performed the right effects.