[WOR-72] Add workspace type field to WORKSPACES table (#1618)
Context

We want to be able to distinguish between those Workspaces managed by Rawls vs those managed by an external system (terra-workspace-manager in this case). To do so, we will need to add a new field to the WORKSPACE table and associated models to allow consuming code to know whether it should delegate operations to the external system.

This PR

Adds a migration that tacks on a workspace_type field to the WORKSPACE table. I have tested this against a clone of DEV and observed no issues during the run of the migration at boot time.