[WOR-129] Add model updates to support workspace type (#1629)
Context

Follow-on to #1618, where we added the workspace_type field to the WORKSPACE table. The next step is to thread the new field through to the related model classes.

This PR

Adds a new WorkspaceType enum to capture states this field can take
Threads this field through the Workspace model and associated calls sites
This field is being hardcoded to rawls to start. The notion is that Azure workspaces created via the updated createWorkspace endpoint will receive a mc value.
Note This PR is being created against [WOR-72] Add workspace type field to WORKSPACES table #1618. I expect this will merge to that PR and then that one will go out as a single unit.