[WOR-73] Initial create MC workspace API and service implementation (#1631)
Context

In #1629 we threaded through the new workspace_type from the DB to the associated models. Now, we need to wire up and conditionally create workspaces in WSM, and potentially related Azure cloud contexts.

This PR

Creates a new MultiCloudWorkspaceService which knows how to provision an azure workspace. For now, this codepath is limited to creating workspaces with Azure cloud contexts, attempts to call with Gcp will result in an exception
Gates access to that method by requiring a new config setting multiCloudWorkspaces.enabled be set to true, and that the incoming request has the cloudPlatform field set
Adds tests for the new service