[AJ-316] Do not log full exception stack trace as this can cause massive log spew (#1641)
Context

We are seeing a tremendous amount of log spew when copying a file during cloning 403s with Google. While we need to figure out the underlying reason and clear out the offending copy operations from the queue, we should also avoid logging the full exception stack as that is pushing us over our logging limits when the copy job retries. The logging message provides enough context without needing a full stack trace.

This PR

Removes the exception from the log statement