[WOR-167] Always end the span in the event of an IOException (#1661)
Context

I'm observing oddities in the tracing results we're getting when updating billing info
After looking into the logs, I'm seeing many SocketReadTimeouts to google associated with this call. Given that, it seems likely that the call to executeGoogleRequest is throwing an IOException and entering into a retry/backoff cycle, which is preventing the final call to span.end, resulting in missed request traces.

This PR

Adds a finally block to ensure we always end the span.