Limit which snapshot we use when testing analysis (#1686)
Context

One of the Snapshot API Spec tests was relying on a single well-known snapshot being present in the list results. A new snapshot was created in that environment. The API returned that snapshot first in the enumeration, causing the test to fail since the schema and data are incorrect.

This PR

Adds a filter for the specific snapshot name that we are looking for