[WOR-187] Add "fake" billing profile manager DAO (#1690)
Context

In order to create Azure workspaces, an azure billing profile (aka billing project in rawls) must be available. Rather than formally introduce the notion of Azure billing information to Rawls, the plan is to have the terra-billing-profile-manager be the source of truth for these records.

Until that service is ready (and to unblock downstream integration + UI efforts for the workspaces team), we'll be introducing a "fake" billing profile manager in Rawls that will serve a hardcoded azure billing profile. Access to this billing profile will be gated by the same SAM permissions required for any other billing profile, as well as membership in the Azure alpha testing group.

This PR

Adds a BillingProfileManagerDAO class that returns a list of billing profiles for the user
Integrates the BillingProfileManagerDAO with the UserService, specifically the listBillingProfilesV2 call. In my testing this was the call that needs modification to get the fake profile into the dropdown on the createWorkspace dialog.
Adds tests for the new DAO + unit level coverage for the existing listBillingProfilesV2 method