Support MC workspace creation (#1742)
Context

We would like the ability to create an azure workspace against the fake azure billing profile setup in WOR-187. To do so, we'll route creation requests from the UI against that billing profile to the MC workspace creation codepath. All other creation requests will be routed to the legacy WorkspaceService codepath.

This PR

Adds a createMultiCloudOrRawlsWorkspace method to the MultiCloudWorkspaceService class. This method checks if the billing profile is the azure billing profile setup in WOR-187. If so, it will attempt to create a workspace using the MC workspace codepath.
Adds a SAM check to the MC workspace creation flow for createWorkspace permissions against the billing profile.
Removes the POST /mc endpoint that we were previously using to create Azure workspaces.
Hardcodes the eastus region. This will need to be threaded in from the UI (see WOR-279)