[WOR-407] Upgrade sentry and emit exceptions in global handler (#1933)
Context

Since we do not log most stack traces to our logs (due to quota management with logit, I believe) we are losing valuable exception information due to our sole reliance on Sentry logback. It would be helpful to have those stack traces for diagnostic purposes.

This PR

Adds Sentry.captureException calls to the Rawls Akka HTTP global exception handler. This will attach useful stack traces to exceptions that are emitted to Sentry.
Upgrades sentry to 6.2.1. The DefaultSentryClientFactory class we were relying on for our custom event filtering is gone. The guidance from sentry is to pass in a callback during Sentry initialization, which I've done here.