[WOR-431] Refactor V2 billing project creation code (#1953)
Context

We need to be able to create azure billing profiles in billing profile manager as part of azure billing project creation. This PR is the first step in the refactoring necessary to introduce that functionality. This step is focused solely on refactoring of existing functionality with an eye toward review-ability and more easily digestible PRs; it does not introduce any new azure/bpm related code.

This PR

Extracts the V2 billing project creation code path to a separate class for improved testability. The current coverage is achieved via heavyweight integration tests on the V2 billing endpoints. I've dubbed this class BillingProjectOrchestrator.
Introduces a billing project repository, again for improved testability. The intention of this class is to isolate all slick DB calls for billing project operations and allow the orchestration layer to test it's business logic without the somewhat elaborate mocking needed for slick data sources or spinning up a real DB connection.