[WOR-431] Optionally create billing profile records and link for azure billing projects (#1961)
Context

(This is the follow up to #1953)
We need to be able to create azure billing profiles in billing profile manager as part of azure billing project creation. This PR is the second step in the refactoring necessary to introduce that functionality. This step introduces the code path necessary to
determine if an incoming billing project creation request is eligible for linkage with a new bililng profile in billing profile manager (i.e, azure billing project creation requests)

This PR

Introduces a BillingProjectCreator trait, of which there are two implementations;
BpmBillingProjectCreator, which will handle the validation and creation of azure billing projects
GoogleBillingProjectCreator which will handle the current GCP codepath
Updates the BillingProfileManagerDAO class with the needed client methods to validate and create azure billing profiles