[WOR-492] Check that a user is enabled before getting a workspace context (#2025)
Context

We are varying the error messaging for disabled users when fetching a workspace depending on whether the workspace exists.

This PR

Adds a check with SAM to determine if the user is disabled when getting a workspace. If so, the get() call fails out with the same error message, regardless of whether the workspace exists.