[PPWM-92] Only provide requester pays project for requester pays buckets (#2106)
RR: https://broadworkbench.atlassian.net/browse/PPWM-92
Part of the issue for workspaces in service perimeters is that we always
provice a requester pays project, even when the bucket is not requester
pays. For AoU and baseline, none of these workspaces are RP and our
requests fail as the provided google project to bill is not whitelisted
for egress.