[WOR-529] Orchestrate Cloning Azure Workspaces (#2142)
RR: https://broadworkbench.atlassian.net/browse/WOR-529
Add `cloneMultiCloudWorkspace` to the `MultiCloudWorkspaceService`. This
verifies that the user has the necessary cloud-agnostic permissions to
clone the source workspace then delegates to the either
`cloneAzureWorkspace` (WOR-625) or the GCP `WorkspaceService` to clone
the google workspace.
I've included a bit of refactoring/flattening the source code to avoid
duplicating work and more code reuse.