[WOR-751] Fail on deletion if the workspace is MC and WSM errors out (#2215)
Context

Rawls eats any exceptions from WSM when deleting workspaces, instead only logging a warning if the errors is not 404. Rawls would then blindly move on and delete the rawls workspace record, orphaning any WSM workspaces and resources.

This PR

When deleting a WSM workspace, we now check if the workspace is of type mc. If so, we error out on any failure from WSM that is not 404 and leave the Rawls workspace record intact. If the workspace is not mc (i.e., a snapshot workspace), the old behavior of eating the exception and moving on is retained.
Moves the WSM workspace deletion call to a position before the deletion of the Rawls workspace record on L725, again so the Rawls workspace is intact and a deletion may be retried.