[WOR-802] Wire up optional LZ ID for testing purposes (#2294)
Context

We want to be able to reuse landing zones in testing scenarios. To this end, I've introduced the notion of attaching landing zones in the Landing Zone Service (and it's owning service WSM). This allows a consumer to provide a landing zone ID and an attach arg which when set will direct the LZS to reuse previously deployed Azure resources rather than creating them fresh and burning 10 minutes to do so. Rawls will need to accommodate this and know that it should provide the LZ ID and attach arg.

This PR

Adds an optional landing zone ID to the azure managed app coords parameter passed in via the billing project creation request. This allows the reuse of the same LZ ID from test to test.
Adds a "landingZoneAllowAttach" configuration variable, defaulting to false. If false and a landing zone parameter is passed in the billing project creation request, an exception is thrown and creation fails. If true, Rawls will pass the "attach -> true" parameter to WSM and LZS.
Removes dead WSM methods that were removed as part of the region updates in WSM
Various scalafmt'ing