[WOR-1035] Wire up workspace deletion for MC workspaces (#2441)
Why

This is the second in the series of PRs to introduce a POST workspace deletion endpoint and to shift over to the async WSM deletion API. See this initial PR and the linked ticket for the broader context. The intention is to rebase and merge this to develop once #2433 merges.

This PR

Shifts the entrypoint for workspace deletion to a new method in the MultiCloudWorkspaceService, deleteMultiCloudOrRawlsWorkspace. The service delegates to the WorkspaceService in the event that the workspace being deleted is a rawls workspace, in a manner very similar to workspace creation.
The MC workspace deletion code in WorkspaceService will be removed in a follow-on PR