[WOR-1188] Include policies in workspace list response, extract workspace aggregation to a separate class (#2491)
Context

We want to include policies in the workspace listing response.

For the existing Rawls getWorkspace codepath, we are making two identical calls to WSM's getWorkspace API: one so we can determine the cloud platform and one to get policies. This is a redundant 100+ ms roundtrip to WSM. For getting a single workspace, not a huge deal. For listing workspaces, this is expensive.

Rather than try to shoehorn a call to the existing getAzureCloudContextFromWorkspaceManager method in WorkspaceService, I elected to perform some refactoring with the goal of making this code more focused and testable, and making the error handling clearer.

To that end, I've introduced the idea of an AggregatedWorkspace. This is the combination of a "rawls" workspace and any other data from upstream services. For now, the only upstream service is Workspace Manager. I've similarly introduced an AggregatedWorkspaceService that isolates the aggregation to a more focused class rather than dumping it into the already huge WorkspaceService class.

The AggregatedWorkspaceService class exposes a single method for now -- getAggregatedWorkspace. This has the WSM connection and stitching logic.

An additional goal was to introduce clearer exception handling. I think the upper layers of Rawls should not have to care about handling WSM client API exceptions -- this is a burden for developers to always have to remember to handle 404, 500s, etc. Instead this new service returns more semantically meaningful exceptions that upper layers can handle as they see fit.

This PR

Implements the refactoring and associated tests as described above
Adds a listWorkspace integration test to the E2E tests.