[WOR-1193] Workspace deletion async runner (#2533)
Context

We are moving to an async model for workspace deletion. In this PR, I implemented the wireup of the API and creation of the monitor record that hands off work to a backend runner task. This PR implements the runner task to orchestrate the async workspace deletion.

This PR

The meat of the PR is in the WorkspaceDeletionRunner class and associated helper classes. This class starts the deletion jobs with downstream systems (Leo + WSM) and the eventual Rawls record deletion.
This is somewhat stateless approach. For each step we adjust the relevant job record in the DB to reflect the current step (i.e., WorkspaceDeletionInit, upon completion, will change the record job type to LeoAppDeletion).
I broke out the Leo + WSM deletion code into their own classes for improved testability and hopefully clearer error handling.
(Finally?) adds a WorkspaceRepository class, again to aid testability. It's contained to basic get/delete + updating state.
I've put this implementation on a BEE here. Happy to add folks to my sandbox billing project for testing purposes.