AJ-1309: Honor the Rawls `WorkspaceType` when creating a WSM workspace just in time as part of creating a Snapshot.
The existing logic blindly creates a workspace without consulting the Rawls `WorkspaceType`, so it will defaulted to having a Rawls stage in WSM even if what is _in_ Rawls is an MC workspace.

Honoring the `WorkspaceType` should help avoid having the data in WSM drift from what is in Rawls going forward as this logic evolves to eventually do a better job supporting Azure.

This is in preparation for making changes to `AggregatedWorkspaceService`, which will start to be used by `SnapshotService` for both GCP and Azure workspaces.