[WOR-1452] Call Leo during Google workspace deletion to cleanup dangling resources (#2745)
Context

We want to notify Leonardo at the time of Google workspace deletion to cleanup any remaining internal Leo DB records and their linked Sam resources. The premise is that Rawls will delete the workspace google cloud project as part of workspace deletion, cleaning up the actual cloud resources, so Leo does not have anything further to do in GCP and only needs to reconcile it's state.

This PR

Adds the needed API to the Leo DAO
Refactors the current LeonardoDeletionAction class to a LeonardoService class. The intention is to keep the Leo DAO "dumb" and let the service class contain the requisite retry and error handling logic. Rawls' current pattern for DAOs is inconsistent. For example, HttpGoogleServicesDAO does all of the above, while the WSM DAO is much lighter weight and pushes the error handling and retries to the callsites.
I'd like to centralize on a pattern of "bare, dumb DAO" + "smarter service that knows how to retry, handle errors and propagate a sane exception hierarchy to upper layers". This is a baby step in that direction.
Removes support for deleting MC azure workspaces via the current WorkspaceService deleteWorkspace codepath. Nothing calls this codepath for MC workspaces anymore and so we shouldn't need the added complexity of the isAzureMcWorkspace checks that are peppered throughout that code.