[WOR-1555] Restore original Google workspace deletion order of operations  (#2762)
Ticket: https://broadworkbench.atlassian.net/browse/WOR-1555

Leo is calling sam as the user's pet to delete resources during the workspace deletion flow. However, with the previous change the google project housing the Pet SA is being deleted before the call to Leo, so there is concern that the leo call will fail out. In my testing I have not been able to reproduce this, but until we can get a firm grasp on the situation I am going to revert the changes to the WorkspaceService to their original flow.

(This is a full revert via the git cli, there were intervening scalafmt commits that caused conflicts which I resolved manually)

This reverts #2745