[WOR-1197] Remove dupes and upgrade liquibase (#3001)
This upgrades liquibase to latest. A previous attempt failed out due to checksum validation failures at runtime. I pulled the changeset checksums from the DATABASECHANGELOG table in dev to a Bee. Running an upgraded liquibase reproduced the checksum errors. There are two issues:
a) string interpolation within the changeset causing the checksum to be different than when the changeset originally ran or
b) re-used id values between changesets

I've added a 'validCheckSum' element of '1:any' to the failing changesets, assuming they will not be changed in the future--all of the offending changesets are quite old and unlikely to change. This will prevent us from having to re-compute checksums any time there is a new algorithm version.