[CORE-671] Make bucket read-only in locked workspaces (#3475)
* Make bucket read-only when locking a workspace

* Restore bucket IAM when unlocking workspace

* Fix compilation and test errors

* Add unit tests

* Scalafmt

* Do not change bucket permissions for Rawls SA

* Fix overriding project owner bucket permissions

A different policy email is used depending on
whether the workspace has an auth domain.

* Fix iamPolicyVersion

If a policy uses v3 and Rawls tries to
override it with a v1 policy google will
throw an error. However if a v3 policy is
set but only a v1 policy is needed Google
will use that one.

* Handle fast pass permissions

When a workspace is created the user
and pet service account emails are
directly given write access to the bucket.
This change removes the grants from a locked
workspace and adds them back on unlock.

* Fix mocking unit tests

* Scalafmt

* Refactor getting policy emails based on auth domain

* Fix checking if auth domain is empty

* Fix checking if auth domain is empty when creating workspace

* Scalafmt

---------

Co-authored-by: Saman Ehsan <sehsan@manifold.ai>