Do not check disk status when locking workspaces
If a user has stopped runtimes/apps the disk
status will be ready. We still want to allow
users to lock their workspaces in this case.
Also if a disk has the creating/restoring/
deleting status that would only happen if an
app or runtime is being created or deleted.